split sender and task modules
