+++ reductstore/src/replication/replication_task.rs
@@ -1,26 +1,29 @@
 // Copyright 2023-2024 ReductStore
 // Licensed under the Business Source License 1.1
 
-use crate::replication::remote_bucket::{create_remote_bucket, RemoteBucket};
-use crate::replication::transaction_filter::TransactionFilter;
-use crate::replication::transaction_log::TransactionLog;
-use crate::replication::{Transaction, TransactionNotification};
-use crate::storage::storage::Storage;
-
-use log::{debug, error, info};
-use reduct_base::error::{ErrorCode, ReductError};
-
-use crate::replication::diagnostics::DiagnosticsCounter;
-use futures_util::stream::Concat;
-use reduct_base::msg::diagnostics::Diagnostics;
-use reduct_base::msg::replication_api::{ReplicationInfo, ReplicationSettings};
+use std::cmp::PartialEq;
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
+
+use futures_util::stream::Concat;
+use log::{debug, error, info};
 use tokio::fs;
 use tokio::sync::RwLock;
 use tokio::time::sleep;
+
+use reduct_base::error::{ErrorCode, ReductError};
+use reduct_base::msg::diagnostics::Diagnostics;
+use reduct_base::msg::replication_api::{ReplicationInfo, ReplicationSettings};
+
+use crate::replication::diagnostics::DiagnosticsCounter;
+use crate::replication::remote_bucket::{create_remote_bucket, RemoteBucket};
+use crate::replication::replication_sender::{ReplicationSender, SyncState};
+use crate::replication::transaction_filter::TransactionFilter;
+use crate::replication::transaction_log::TransactionLog;
+use crate::replication::{Transaction, TransactionNotification};
+use crate::storage::storage::Storage;
 
 pub struct ReplicationTask {
     name: String,
@@ -112,20 +115,42 @@ impl ReplicationTask {
                 error!("Failed to initialize transaction logs: {:?}", err);
             }
 
-            let sender = ReplicationSender {
-                replication_name,
-                log_map: thr_log_map,
-                storage: thr_storage,
-                config,
-                hourly_diagnostics: thr_hourly_diagnostics,
-                bucket: thr_bucket,
-                remote_bucket_timeout: Duration::from_secs(5),
-            };
+            let sender = ReplicationSender::new(
+                replication_name.clone(),
+                thr_log_map,
+                thr_storage.clone(),
+                config.clone(),
+                thr_hourly_diagnostics,
+                thr_bucket,
+            );
 
             loop {
-                // NOTE: we don't want to spin the CPU when there is nothing to do
-                if sender.run().await {
-                    sleep(Duration::from_millis(100)).await;
+                match sender.run().await {
+                    SyncState::SyncedOrRemoved => {}
+                    SyncState::NotAvailable => {
+                        sleep(Duration::from_secs(5)).await;
+                    }
+                    SyncState::NoTransactions => {
+                        // NOTE: we don't want to spin the CPU when there is nothing to do or the bucket is not available
+                        sleep(Duration::from_millis(50)).await;
+                    }
+                    SyncState::BrokenLog(entry_name) => {
+                        info!("Transaction log is corrupted, dropping the whole log");
+                        let path = ReplicationTask::build_path_to_transaction_log(
+                            thr_storage.read().await.data_path(),
+                            &config.src_bucket,
+                            &entry_name,
+                            &replication_name,
+                        );
+                        if let Err(err) = fs::remove_file(&path).await {
+                            error!("Failed to remove transaction log: {:?}", err);
+                        }
+
+                        info!("Creating a new transaction log");
+                        TransactionLog::try_load_or_create(path, TRANSACTION_LOG_SIZE)
+                            .await
+                            .unwrap();
+                    }
                 }
             }
         });
@@ -236,135 +261,22 @@ impl ReplicationTask {
     }
 }
 
-/// Internal worker for replication to process a sole iteration of the replication loop.
-struct ReplicationSender {
-    replication_name: String,
-    log_map: Arc<RwLock<HashMap<String, RwLock<TransactionLog>>>>,
-    storage: Arc<RwLock<Storage>>,
-    config: ReplicationSettings,
-    hourly_diagnostics: Arc<RwLock<DiagnosticsCounter>>,
-    bucket: Arc<RwLock<dyn RemoteBucket + Send + Sync>>,
-    remote_bucket_timeout: Duration,
-}
-
-impl ReplicationSender {
-    pub async fn run(&self) -> bool {
-        let mut no_transactions = true;
-        for (entry_name, log) in self.log_map.read().await.iter() {
-            let tr = log.write().await.front().await;
-            match tr {
-                Ok(None) => {}
-                Ok(Some(transaction)) => {
-                    no_transactions = false;
-
-                    debug!("Replicating transaction {}/{:?}", entry_name, transaction);
-
-                    let read_record_from_storage = async {
-                        let mut atempts = 3;
-                        loop {
-                            let read_record = async {
-                                self.storage
-                                    .read()
-                                    .await
-                                    .get_bucket(&self.config.src_bucket)?
-                                    .begin_read(&entry_name, *transaction.timestamp())
-                                    .await
-                            };
-                            let record = read_record.await;
-                            match record {
-                                Err(ReductError {
-                                    status: ErrorCode::TooEarly,
-                                    ..
-                                }) => {
-                                    debug!("Transaction is too early, retrying later");
-                                    sleep(Duration::from_millis(10)).await;
-                                    atempts -= 1;
-                                }
-
-                                _ => {
-                                    atempts = 0;
-                                }
-                            }
-
-                            if atempts == 0 {
-                                break record;
-                            }
-                        }
-                    };
-
-                    let record_to_sync = match read_record_from_storage.await {
-                        Ok(record) => record,
-                        Err(err) => {
-                            log.write().await.pop_front().await.unwrap();
-                            error!("Failed to read record: {}", err);
-                            self.hourly_diagnostics.write().await.count(Err(err));
-                            break;
-                        }
-                    };
-
-                    let mut bucket = self.bucket.write().await;
-                    match bucket.write_record(entry_name, record_to_sync).await {
-                        Ok(_) => {
-                            if bucket.is_active() {
-                                self.hourly_diagnostics.write().await.count(Ok(()));
-                            }
-                        }
-                        Err(err) => {
-                            debug!("Failed to replicate transaction: {:?}", err);
-                            self.hourly_diagnostics.write().await.count(Err(err));
-                        }
-                    }
-
-                    if bucket.is_active() {
-                        log.write().await.pop_front().await.unwrap();
-                    } else {
-                        // We keep transactions if the destination bucket is not available
-                        drop(bucket); // drop lock
-                        sleep(self.remote_bucket_timeout).await; // FIXME: this is a temporary solution
-                    }
-                }
-
-                Err(err) => {
-                    error!("Failed to read transaction: {:?}", err);
-                    info!("Transaction log is corrupted, dropping the whole log");
-                    let mut log = log.write().await;
-
-                    let path = ReplicationTask::build_path_to_transaction_log(
-                        self.storage.read().await.data_path(),
-                        &self.config.src_bucket,
-                        &entry_name,
-                        &self.replication_name,
-                    );
-                    if let Err(err) = fs::remove_file(&path).await {
-                        error!("Failed to remove transaction log: {:?}", err);
-                    }
-
-                    info!("Creating a new transaction log");
-                    *log = TransactionLog::try_load_or_create(path, TRANSACTION_LOG_SIZE)
-                        .await
-                        .unwrap();
-                }
-            }
-        }
-
-        return no_transactions;
-    }
-}
-
+#[cfg(target_os = "linux")]
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::replication::Transaction;
-    use crate::storage::bucket::RecordReader;
     use async_trait::async_trait;
     use bytes::Bytes;
-
     use mockall::mock;
-    use reduct_base::error::ErrorCode;
-    use reduct_base::msg::bucket_api::BucketSettings;
-    use reduct_base::msg::diagnostics::{DiagnosticsError, DiagnosticsItem};
-    use reduct_base::Labels;
     use rstest::*;
+
+    use reduct_base::msg::bucket_api::BucketSettings;
+    use reduct_base::msg::diagnostics::DiagnosticsItem;
+    use reduct_base::Labels;
+
+    use crate::replication::Transaction;
+    use crate::storage::bucket::RecordReader;
+
+    use super::*;
 
     mock! {
         RmBucket {}
@@ -382,408 +294,92 @@ mod tests {
 
     }
 
-    #[cfg(target_os = "linux")]
-    mod replication_task {
-        use super::*;
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_transaction_log_init(
-            remote_bucket: MockRmBucket,
-            settings: ReplicationSettings,
-        ) {
-            let replication = build_replication(remote_bucket, settings).await;
-            assert_eq!(replication.log_map.read().await.len(), 1);
-            assert!(
-                replication.log_map.read().await.contains_key("test"),
-                "Transaction log is initialized"
-            );
-        }
-
-        /// TODO check notify method
-
-        async fn build_replication(
-            remote_bucket: MockRmBucket,
-            settings: ReplicationSettings,
-        ) -> ReplicationTask {
-            let tmp_dir = tempfile::tempdir().unwrap().into_path();
-
-            let filter = TransactionFilter::new(
-                settings.src_bucket.clone(),
-                settings.entries.clone(),
-                settings.include.clone(),
-                settings.exclude.clone(),
-            );
-
-            let storage = Arc::new(RwLock::new(Storage::load(tmp_dir, None).await));
-
-            {
-                let mut lock = storage.write().await;
-
-                let bucket = lock
-                    .create_bucket("src", BucketSettings::default())
-                    .unwrap();
-
-                let tx = bucket
-                    .write_record("test", 10, 4, "text/plain".to_string(), Labels::new())
-                    .await
-                    .unwrap();
-                tx.send(Ok(Some(Bytes::from("test")))).await.unwrap();
-                tx.send(Ok(None)).await.unwrap_or(());
-            }
-
-            let repl = ReplicationTask::build(
-                "test".to_string(),
-                settings,
-                Arc::new(RwLock::new(remote_bucket)),
-                filter,
-                storage,
-            );
-
-            sleep(Duration::from_millis(5)).await; // wait for the transaction log to be initialized in worker
-            repl
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_replication_ok_active(
-            mut remote_bucket: MockRmBucket,
-            notification: TransactionNotification,
-            settings: ReplicationSettings,
-        ) {
-            remote_bucket.expect_write_record().returning(|_, _| Ok(()));
-            remote_bucket.expect_is_active().return_const(true);
-            let replication = build_replication(remote_bucket, settings).await;
-
-            replication.notify(notification).await.unwrap();
-            assert!(transaction_log_is_empty(&replication).await);
-            assert_eq!(
-                replication.info().await,
-                ReplicationInfo {
-                    name: "test".to_string(),
-                    is_active: true,
-                    is_provisioned: false,
-                    pending_records: 0,
-                }
-            );
-            assert_eq!(
-                replication.diagnostics().await,
-                Diagnostics {
-                    hourly: DiagnosticsItem {
-                        ok: 60,
-                        errored: 0,
-                        errors: HashMap::new(),
-                    }
-                }
-            )
-        }
+    #[rstest]
+    #[tokio::test]
+    async fn test_transaction_log_init(remote_bucket: MockRmBucket, settings: ReplicationSettings) {
+        let replication = build_replication(remote_bucket, settings).await;
+        assert_eq!(replication.log_map.read().await.len(), 1);
+        assert!(
+            replication.log_map.read().await.contains_key("test"),
+            "Transaction log is initialized"
+        );
     }
 
-    mod test_replication_sender {
-        use super::*;
+    async fn build_replication(
+        remote_bucket: MockRmBucket,
+        settings: ReplicationSettings,
+    ) -> ReplicationTask {
+        let tmp_dir = tempfile::tempdir().unwrap().into_path();
 
-        #[rstest]
-        #[tokio::test]
-        async fn test_replication_ok_not_active(
-            mut remote_bucket: MockRmBucket,
-            settings: ReplicationSettings,
-        ) {
-            remote_bucket.expect_write_record().returning(|_, _| Ok(()));
-            remote_bucket.expect_is_active().return_const(false);
-            let sender = build_sender(remote_bucket, settings).await;
+        let filter = TransactionFilter::new(
+            settings.src_bucket.clone(),
+            settings.entries.clone(),
+            settings.include.clone(),
+            settings.exclude.clone(),
+        );
 
-            let transaction = Transaction::WriteRecord(10);
-            imitate_write_record(&sender, &transaction).await;
+        let storage = Arc::new(RwLock::new(Storage::load(tmp_dir, None).await));
 
-            assert!(!sender.run().await);
-            assert_eq!(
-                sender
-                    .log_map
-                    .read()
-                    .await
-                    .get("test")
-                    .unwrap()
-                    .read()
-                    .await
-                    .front()
-                    .await,
-                Ok(Some(transaction)),
-            );
+        {
+            let mut lock = storage.write().await;
 
-            assert_eq!(
-                sender.hourly_diagnostics.read().await.diagnostics(),
-                DiagnosticsItem {
-                    ok: 0,
-                    errored: 0,
-                    errors: HashMap::new(),
-                },
-                "should not count errors for non active replication"
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_replication_comm_err(mut remote_bucket: MockRmBucket) {
-            remote_bucket
-                .expect_write_record()
-                .returning(|_, _| Err(ReductError::new(ErrorCode::Timeout, "Timeout")));
-            remote_bucket.expect_is_active().return_const(false);
-            let sender = build_sender(remote_bucket, settings()).await;
-
-            let transaction = Transaction::WriteRecord(10);
-            imitate_write_record(&sender, &transaction).await;
-
-            assert!(!sender.run().await);
-
-            assert_eq!(
-                sender
-                    .log_map
-                    .read()
-                    .await
-                    .get("test")
-                    .unwrap()
-                    .read()
-                    .await
-                    .front()
-                    .await,
-                Ok(Some(transaction)),
-            );
-
-            let diagnostics = sender.hourly_diagnostics.read().await.diagnostics();
-            assert_eq!(diagnostics.ok, 0);
-            assert!(
-                diagnostics.errored > 0,
-                "We have at least one errored transaction",
-            );
-            assert!(
-                diagnostics.errors[&-2].count > 0,
-                "We have at least one timeout error",
-            );
-            assert!(
-                diagnostics.errors[&-2].last_message.contains("Timeout"),
-                "We have at least one timeout error",
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_replication_not_found(
-            mut remote_bucket: MockRmBucket,
-            settings: ReplicationSettings,
-        ) {
-            remote_bucket.expect_write_record().returning(|_, _| Ok(()));
-            remote_bucket.expect_is_active().return_const(true);
-            let sender = build_sender(remote_bucket, settings).await;
-
-            let transaction = Transaction::WriteRecord(10);
-            imitate_write_record(&sender, &transaction).await;
-            sender
-                .storage
-                .write()
-                .await
-                .get_mut_bucket("src")
-                .unwrap()
-                .remove_entry("test")
-                .unwrap();
-
-            assert!(!sender.run().await);
-            assert!(
-                sender
-                    .log_map
-                    .read()
-                    .await
-                    .get("test")
-                    .unwrap()
-                    .read()
-                    .await
-                    .is_empty(),
-                "We don't keep the transaction for a non existing record"
-            );
-
-            let diagnostics = sender.hourly_diagnostics.read().await.diagnostics();
-            assert_eq!(
-                diagnostics,
-                DiagnosticsItem {
-                    ok: 0,
-                    errored: 60,
-                    errors: HashMap::from_iter(vec![(
-                        404,
-                        DiagnosticsError {
-                            count: 1,
-                            last_message: "Entry 'test' not found in bucket 'src'".to_string(),
-                        }
-                    )]),
-                }
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_replication_too_early_ok(
-            mut remote_bucket: MockRmBucket,
-            settings: ReplicationSettings,
-        ) {
-            remote_bucket.expect_write_record().returning(|_, _| Ok(()));
-            remote_bucket.expect_is_active().return_const(true);
-            let sender = build_sender(remote_bucket, settings).await;
-
-            sender
-                .log_map
-                .read()
-                .await
-                .get("test")
-                .unwrap()
-                .write()
-                .await
-                .push_back(Transaction::WriteRecord(20))
-                .await
-                .unwrap();
-
-            let tx = sender
-                .storage
-                .write()
-                .await
+            let bucket = lock
                 .create_bucket("src", BucketSettings::default())
-                .unwrap()
-                .write_record("test", 20, 4, "".to_string(), Labels::new())
-                .await
                 .unwrap();
 
-            let hourly_diagnostics = sender.hourly_diagnostics.clone();
+            let tx = bucket
+                .write_record("test", 10, 4, "text/plain".to_string(), Labels::new())
+                .await
+                .unwrap();
+            tx.send(Ok(Some(Bytes::from("test")))).await.unwrap();
+            tx.send(Ok(None)).await.unwrap_or(());
+        }
 
-            tokio::spawn(async move {
-                // we need to spawn a task to check the state in the attempt loop
-                sender.run().await
-            });
+        let repl = ReplicationTask::build(
+            "test".to_string(),
+            settings,
+            Arc::new(RwLock::new(remote_bucket)),
+            filter,
+            storage,
+        );
 
-            tx.send(Ok(Some(Bytes::from("xxxx")))).await.unwrap();
-            sleep(Duration::from_millis(15)).await;
+        sleep(Duration::from_millis(5)).await; // wait for the transaction log to be initialized in worker
+        repl
+    }
 
-            let diagnostics = hourly_diagnostics.read().await.diagnostics();
-            assert_eq!(
-                diagnostics,
-                DiagnosticsItem {
+    #[rstest]
+    #[tokio::test]
+    async fn test_replication_ok_active(
+        mut remote_bucket: MockRmBucket,
+        notification: TransactionNotification,
+        settings: ReplicationSettings,
+    ) {
+        remote_bucket.expect_write_record().returning(|_, _| Ok(()));
+        remote_bucket.expect_is_active().return_const(true);
+        let replication = build_replication(remote_bucket, settings).await;
+
+        replication.notify(notification).await.unwrap();
+        sleep(Duration::from_millis(50)).await;
+        assert!(transaction_log_is_empty(&replication).await);
+        assert_eq!(
+            replication.info().await,
+            ReplicationInfo {
+                name: "test".to_string(),
+                is_active: true,
+                is_provisioned: false,
+                pending_records: 0,
+            }
+        );
+        assert_eq!(
+            replication.diagnostics().await,
+            Diagnostics {
+                hourly: DiagnosticsItem {
                     ok: 60,
                     errored: 0,
                     errors: HashMap::new(),
                 }
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_replication_too_early_err(
-            mut remote_bucket: MockRmBucket,
-            settings: ReplicationSettings,
-        ) {
-            remote_bucket.expect_write_record().returning(|_, _| Ok(()));
-            remote_bucket.expect_is_active().return_const(true);
-            let sender = build_sender(remote_bucket, settings).await;
-
-            sender
-                .log_map
-                .read()
-                .await
-                .get("test")
-                .unwrap()
-                .write()
-                .await
-                .push_back(Transaction::WriteRecord(20))
-                .await
-                .unwrap();
-
-            let _tx = sender
-                .storage
-                .write()
-                .await
-                .create_bucket("src", BucketSettings::default())
-                .unwrap()
-                .write_record("test", 20, 4, "".to_string(), Labels::new())
-                .await
-                .unwrap();
-
-            sender.run().await;
-
-            let diagnostics = sender.hourly_diagnostics.read().await.diagnostics();
-            assert_eq!(
-                diagnostics,
-                DiagnosticsItem {
-                    ok: 0,
-                    errored: 20,
-                    errors: HashMap::from_iter(vec![(
-                        425,
-                        DiagnosticsError {
-                            count: 1,
-                            last_message: "Record with timestamp 20 is still being written"
-                                .to_string()
-                        }
-                    )]),
-                }
-            );
-        }
-
-        async fn imitate_write_record(sender: &ReplicationSender, transaction: &Transaction) {
-            sender
-                .log_map
-                .read()
-                .await
-                .get("test")
-                .unwrap()
-                .write()
-                .await
-                .push_back(transaction.clone())
-                .await
-                .unwrap();
-            let tx = sender
-                .storage
-                .write()
-                .await
-                .create_bucket("src", BucketSettings::default())
-                .unwrap()
-                .write_record(
-                    "test",
-                    transaction.timestamp().clone(),
-                    4,
-                    "text/plain".to_string(),
-                    Labels::new(),
-                )
-                .await
-                .unwrap();
-            tx.send(Ok(Some(Bytes::from("xxxx")))).await.unwrap();
-        }
-
-        async fn build_sender(
-            remote_bucket: MockRmBucket,
-            settings: ReplicationSettings,
-        ) -> ReplicationSender {
-            let tmp_dir = tempfile::tempdir().unwrap().into_path();
-
-            let storage = Arc::new(RwLock::new(Storage::load(tmp_dir, None).await));
-
-            let log_map = Arc::new(RwLock::new(HashMap::new()));
-            let log = RwLock::new(
-                TransactionLog::try_load_or_create(
-                    storage.read().await.data_path().join("test.log"),
-                    TRANSACTION_LOG_SIZE,
-                )
-                .await
-                .unwrap(),
-            );
-
-            log_map.write().await.insert("test".to_string(), log);
-
-            ReplicationSender {
-                replication_name: "test".to_string(),
-                log_map,
-                storage,
-                config: settings,
-                hourly_diagnostics: Arc::new(RwLock::new(DiagnosticsCounter::new(
-                    Duration::from_secs(1),
-                ))),
-                bucket: Arc::new(RwLock::new(remote_bucket)),
-                remote_bucket_timeout: Duration::from_millis(10),
             }
-        }
+        )
     }
 
     #[fixture]
